Move profiling metrics to tools and rename gdbutil
