+++ README
@@ -354,9 +354,9 @@ default build uses qspinlocks.
 GDB UTILITIES
 -------------
 Helper macros for inspecting x86 descriptor tables live under
-``tools/gdbutil``.  Load them in GDB with::
+``tools/gdbutil.py``.  Load them in GDB with::
 
-    source tools/gdbutil
+    source tools/gdbutil.py
 
 Use ``printdesc`` or ``printdescs`` to display a single descriptor or a
 range of entries.
