+++ meson.build
@@ -225,7 +225,7 @@ fsimg = custom_target('fs.img',
   depends: uprogs_targets,
   build_by_default: true)
 
-executable('phoenix_prof', 'tools/profiling/phoenix_metrics.c',
+executable('phoenix_prof', 'tools/phoenix_metrics.c',
            include_directories: include_directories('.'),
            install: false,
            c_args: common_cargs + ['-DPHOENIX_METRICS_MAIN'])
